fix potentially circular class method call
